put more things in their right place
