+++ src/debug/gui.rs
@@ -1,4 +1,4 @@
-use crate::{gui::GuiComponent, GameRunner};
+use crate::{settings::gui::GuiComponent, GameRunner};
 use egui::{Context, Slider, Window};
 
 #[derive(Hash, PartialEq, Eq)]
