+++ src/gui/mod.rs
@@ -1,8 +1,8 @@
-use crate::GameRunner;
+use crate::{settings::gui::Gui, GameRunner};
 use egui::{ClippedPrimitive, Context, TexturesDelta};
 use egui_wgpu::renderer::{Renderer, ScreenDescriptor};
 use pixels::{wgpu, PixelsContext};
-use winit::{event::VirtualKeyCode, event_loop::EventLoopWindowTarget, window::Window};
+use winit::{event_loop::EventLoopWindowTarget, window::Window};
 
 /// Manages all state required for rendering egui over `Pixels`.
 pub struct Framework {
@@ -47,7 +47,7 @@ impl Framework {
             renderer,
             paint_jobs: Vec::new(),
             textures: TexturesDelta::default(),
-            gui: Gui::new(),
+            gui: Default::default(),
         }
     }
 
@@ -133,89 +133,6 @@ impl Framework {
         let textures = std::mem::take(&mut self.textures);
         for id in &textures.free {
             self.renderer.free_texture(id);
-        }
-    }
-}
-
-pub trait GuiComponent {
-    fn handle_event(&mut self, event: &winit::event::WindowEvent, game_runner: &mut GameRunner);
-    fn ui(
-        &mut self,
-        ctx: &Context,
-        game_runner: &mut GameRunner,
-        ui_visible: bool,
-        is_open: &mut bool,
-    );
-    fn name(&self) -> String;
-}
-
-pub struct SettingsContainer {
-    open: bool,
-    component: Box<dyn GuiComponent>,
-}
-
-impl SettingsContainer {
-    pub fn new(component: Box<dyn GuiComponent>) -> Self {
-        Self {
-            open: false,
-            component,
-        }
-    }
-}
-
-pub struct Gui {
-    // State for the demo app.
-    visible: bool,
-    settings: Vec<SettingsContainer>,
-}
-
-impl Gui {
-    fn new() -> Self {
-        Self {
-            visible: false,
-            settings: vec![],
-        }
-    }
-
-    pub fn add_settings(&mut self, component: Box<dyn GuiComponent>) {
-        self.settings.push(SettingsContainer::new(component));
-    }
-
-    fn handle_event(&mut self, event: &winit::event::WindowEvent, game_runner: &mut GameRunner) {
-        if let winit::event::WindowEvent::KeyboardInput { input, .. } = event {
-            if let Some(code) = input.virtual_keycode {
-                if input.state == winit::event::ElementState::Pressed
-                    && code == VirtualKeyCode::Escape
-                {
-                    self.visible = !self.visible;
-                }
-            }
-        }
-        for g in &mut self.settings {
-            g.component.handle_event(event, game_runner);
-        }
-    }
-
-    fn ui(&mut self, ctx: &Context, game_runner: &mut GameRunner) {
-        if self.visible {
-            egui::TopBottomPanel::top("menubar_container").show(ctx, |ui| {
-                egui::menu::bar(ui, |ui| {
-                    ui.menu_button("Settings", |ui| {
-                        for setting in &mut self.settings {
-                            if ui.button(setting.component.name()).clicked() {
-                                setting.open = !setting.open;
-                                ui.close_menu();
-                            }
-                        }
-                    })
-                });
-            });
-        }
-
-        for setting in &mut self.settings {
-            setting
-                .component
-                .ui(ctx, game_runner, self.visible, &mut setting.open);
         }
     }
 }
